refactor: split monthly cash flow components
